% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_data_Sheet1_GDP_inflation.R
\docType{data}
\name{forecast_data_Sheet1_GDP_inflation}
\alias{forecast_data_Sheet1_GDP_inflation}
\title{Italy forecast data, Sheet1, GDP_inflation indicator}
\format{
\subsection{\code{forecast_data_Sheet1_GDP_inflation}}{

A tibble with 16 rows and 4 columns:
\describe{
\item{entity}{The entity (European Commission, IMF, OECD, ECB)}
\item{period}{The moment when the forecast is revised}
\item{year}{The year of the forecast}
\item{value}{The value of the forecast for the given year}
}
}
}
\source{
extdata/forecasts-italy.xlsx
}
\usage{
forecast_data_Sheet1_GDP_inflation
}
\description{
This is the Italy forecast data, Sheet1 of the forecasts-italy.xlsx file from
December 2023, GDP_inflation indicator, that is included to build an example
dashboard for the NSI scenario.
}
\keyword{datasets}
