% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_data_Sheet3.R
\docType{data}
\name{forecast_data_Sheet3}
\alias{forecast_data_Sheet3}
\title{Eurostatistics forecast data, Sheet3}
\format{
\subsection{\code{forecast_data_Sheet3}}{

A data frame with 21 rows and 9 columns:
\describe{
\item{dim}{The dimension}
\item{European Commission - Summer 2023}{Forecast value from the European
Commission in Summer 2023}
\item{European Commission - Autumn 2023}{Forecast value from the European
Commission in Autumn 2023}
\item{IMF - July 2023}{Forecast value from the IMF in July 2023}
\item{IMF - October 2023}{Forecast value from the IMF in October 2023}
\item{OECD - September 2023}{Forecast value from the OECD in September 2023}
\item{OECD - November 2023}{Forecast value from the OECD in November 2023}
\item{ECB - June 2023}{Forecast value from the ECB in June 2023}
\item{ECB - September 2023}{Forecast value from the ECB in September 2023}
}
}
}
\source{
extdata/forecasts.xlsx
}
\usage{
forecast_data_Sheet3
}
\description{
This is the Eurostatistics forecast data, Sheet3 of the forecasts.xlsx file from
December 2023, representing the forecast for the inflation indicator, that
is included to build an example dashboard for the Eurostatistics scenario.
}
\keyword{datasets}
