% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisco_countries.R
\docType{data}
\name{gisco_countries}
\alias{gisco_countries}
\title{Gisco countries SpatialPolygonsDataFrame}
\format{
An object of class \code{SpatialPolygonsDataFrame} with 257 rows and 12 columns.
}
\source{
https://cran.r-project.org/web/packages/giscoR/giscoR.pdf
}
\usage{
gisco_countries
}
\description{
This is the countries SpatialPolygonsDataFrame, included to draw the polygons
for the all the countries in the maps. It is obtained from the giscoR package.
}
\details{
Generation process:

install.packages("giscoR")

install.packages("sf")

install.packages("sp")

library(giscoR)

library(sf)

library(sp)
}
\section{Get countries borders}{
borders <- giscoR::gisco_get_countries(epsg = "4326", year = "2020", spatialtype = "RG", resolution = "60")

gisco_countries <- sf::as_Spatial(sf::st_geometry(borders), IDs = as.character(1:nrow(borders)))
}

\section{Grab the data from the sf object}{
df <- borders

df$geometry <- NULL

df <- as.data.frame(df)
}

\section{Create the SpatialPolygonsDataFrame}{
gisco_countries <- sp::SpatialPolygonsDataFrame(gisco_countries, data = df)

use_data(gisco_countries, overwrite = TRUE)
}

\keyword{datasets}
