% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_to_default.R
\name{reset_to_default}
\alias{reset_to_default}
\title{Reset the package to its default version}
\usage{
reset_to_default()
}
\value{
Delete all the user package files
}
\description{
Delete user package data and configuration, to reset the package to its
default files and settings.
}
