% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Eurostatistics indicators data}
\format{
\subsection{\code{sample_data}}{

A data frame with 370,658 rows and 5 columns:
\describe{
\item{date}{Date}
\item{IND}{Indicator code}
\item{FREQ}{Indicator frequency: M for monthly, Q for quarterly, A for annually}
\item{DIM}{Dimension code}
\item{obsValue}{Observed value for the indicator}
}
}
}
\source{
extdata/data.csv
}
\usage{
sample_data
}
\description{
This is the Eurostatistics indicators data from December 2023, that is
included to build an example dashboard.
}
\keyword{datasets}
