% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data_italy.R
\docType{data}
\name{sample_data_italy}
\alias{sample_data_italy}
\title{Italy indicators data}
\format{
\subsection{\code{sample_data_italy}}{

A data frame with 2,296 rows and 5 columns:
\describe{
\item{date}{Date}
\item{IND}{Indicator code}
\item{FREQ}{Indicator frequency: M for monthly, Q for quarterly, A for annually}
\item{DIM}{Dimension code}
\item{obsValue}{Observed value for the indicator}
}
}
}
\source{
extdata/data-italy.csv
}
\usage{
sample_data_italy
}
\description{
This is the Italy indicators data from December 2023, that is
included to build an example dashboard for the NSI scenario.
}
\keyword{datasets}
