% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_dim_labels.R
\docType{data}
\name{sample_dim_labels}
\alias{sample_dim_labels}
\title{Eurostatistics dimension codes and their corresponding labels}
\format{
\subsection{\code{sample_dim_labels}}{

A data frame with 38 rows and 3 columns:
\describe{
\item{code}{The dimension code}
\item{label}{The dimension label}
\item{group}{The dimension group, used to group labels in the plots}
}
}
}
\source{
extdata/dim-labels.csv
}
\usage{
sample_dim_labels
}
\description{
This is the Eurostatistics dimension dictionary, for translating dimension
codes to their corresponding labels, from the dim-labels.csv file, included
to build an example dashboard for the Eurostatistics scenario.
}
\keyword{datasets}
