% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_dim_labels_italy.R
\docType{data}
\name{sample_dim_labels_italy}
\alias{sample_dim_labels_italy}
\title{Italy dimension codes and their corresponding labels}
\format{
\subsection{\code{sample_dim_labels_italy}}{

A data frame with 40 rows and 2 columns:
\describe{
\item{code}{The dimension code}
\item{label}{The dimension label}
}
}
}
\source{
extdata/dim-labels-italy.csv
}
\usage{
sample_dim_labels_italy
}
\description{
This is the Italy dimensions dictionary, for translating dimension codes to
their corresponding labels, from the dim-labels-italy.csv file, included
to build an example dashboard for the NSI scenario.
}
\keyword{datasets}
