% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_colors_palette.R
\name{update_colors_palette}
\alias{update_colors_palette}
\title{Update the colors of the lines / bars in the graphs.}
\usage{
update_colors_palette(colors_palette = NULL, bar_chart_colors_palette = NULL)
}
\arguments{
\item{colors_palette}{The color palette for the line plots.}

\item{bar_chart_colors_palette}{The color palette for the bar charts.}
}
\value{
Change the colors_palette in the configuration file.
}
\description{
Update the colors of the lines / bars in the graphs.
}
