% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_dashboard_colors.R
\name{update_dashboard_colors}
\alias{update_dashboard_colors}
\title{Update the dashboard colors}
\usage{
update_dashboard_colors(
  primary_background_color = "#004494",
  primary_color = "white",
  secondary_background_color = "#FFD617",
  secondary_color = "black",
  navbar_border_color = "#002F67",
  storyboard_background_color = "#BFD0E4",
  storyboard_color = "black",
  storyboard_border_color = "#FBC11D",
  on_hover_storyboard_background_color = "#7FA1C9",
  on_hover_storyboard_color = "white",
  overview_text_background_color = "white",
  overview_text_color = "black",
  latest_changes_summary_background_color = "#404040",
  latest_changes_summary_color = "white",
  valuebox1_background_color = "#006FB4",
  valuebox1_color = "white",
  valuebox2_background_color = "#F29527",
  valuebox2_color = "black",
  valuebox3_background_color = "#467A39",
  valuebox3_color = "white",
  description_background_color = "#EBEBEB",
  description_color = "black",
  mobile_chart_title_background_color = "#F2F5F9",
  mobile_chart_title_color = "#004494"
)
}
\arguments{
\item{primary_background_color}{The background color for the navbar.}

\item{primary_color}{The text color for the navbar.}

\item{secondary_background_color}{The background color when hovering or
selecting an item from the main menu.}

\item{secondary_color}{The text color when hovering or selecting an item
from the main menu.}

\item{navbar_border_color}{The border color for the navbar.}

\item{storyboard_background_color}{The background color for the unhovered
items from the storyboard menu.}

\item{storyboard_color}{The text color for the unhovered items from the
storyboard menu.}

\item{storyboard_border_color}{The border color for the selected items from
the storyboard menu.}

\item{on_hover_storyboard_background_color}{The background color when
hovering items from the storyboard menu.}

\item{on_hover_storyboard_color}{The text color when hovering items from
the storyboard menu.}

\item{overview_text_background_color}{The background color for the overview
text (first box appearing in the overview section).}

\item{overview_text_color}{The text color for the overview text (first box
appearing in the overview section).}

\item{latest_changes_summary_background_color}{The background color for the
latest changes summary (second box appearing in the overview section).}

\item{latest_changes_summary_color}{The text color for the latest changes
summary (second box appearing in the overview section).}

\item{valuebox1_background_color}{The background color for the first value
box in the overview section.}

\item{valuebox1_color}{The text color for the first value box in the
overview section.}

\item{valuebox2_background_color}{The background color for the second value
box in the overview section.}

\item{valuebox2_color}{The text color for the second value box in the
overview section.}

\item{valuebox3_background_color}{The background color for the third value
box in the overview section.}

\item{valuebox3_color}{The text color for the third value box in the
overview section.}

\item{description_background_color}{The background color for the description
box on the right side.}

\item{description_color}{The text color for the description box on the right
side.}

\item{mobile_chart_title_background_color}{The background color for the
secondary menu items in the mobile version, when the storyboard is disabled.}

\item{mobile_chart_title_color}{The text color for the secondary menu items
in the mobile version, when the storyboard is disabled.}
}
\value{
Save the new colors.css in the user memory.
}
\description{
Update the dashboard colors
}
