% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_dashboard_structure.R
\name{update_dashboard_structure}
\alias{update_dashboard_structure}
\title{Update the dashboard structure}
\usage{
update_dashboard_structure(file_path)
}
\arguments{
\item{file_path}{The file path of the dashboard structure file.}
}
\value{
Save the new dashboard-structure.xlsx in the user memory.
}
\description{
Update the dashboard structure
}
