% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data.R
\name{update_data}
\alias{update_data}
\title{Update the data}
\usage{
update_data(file_path)
}
\arguments{
\item{file_path}{The file path of the data file.}
}
\value{
Save the new data.csv in the user memory.
}
\description{
Update the data
}
