% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_date_from.R
\name{update_date_from}
\alias{update_date_from}
\title{Update the starting date for the plots}
\usage{
update_date_from(date)
}
\arguments{
\item{date}{The starting date, in "yyyy-mm-dd" format.}
}
\value{
Change the starting date in the configuration file.
}
\description{
Update the starting date for the plots
}
