% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_dim_labels.R
\name{update_dim_labels}
\alias{update_dim_labels}
\title{Update the dim labels}
\usage{
update_dim_labels(file_path)
}
\arguments{
\item{file_path}{The file path of the dim labels file.}
}
\value{
Save the new dim-labels.csv in the user memory.
}
\description{
Update the dim labels
}
