% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_edition.R
\name{update_edition}
\alias{update_edition}
\title{Update the edition of the dashboard}
\usage{
update_edition(edition)
}
\arguments{
\item{edition}{The dashboard edition.}
}
\value{
Change the edition in the configuration file.
}
\description{
Update the edition of the dashboard
}
