% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_footer.R
\name{update_footer}
\alias{update_footer}
\title{Update the footer of the dashboard, appearing in the bottom of the dashboard}
\usage{
update_footer(
  last_update = NULL,
  cookies_policy_link = NULL,
  privacy_policy_link = NULL,
  legal_notice_link = NULL,
  feedback_email = NULL,
  feedback_email_subject = NULL
)
}
\arguments{
\item{last_update}{The last update text.}

\item{cookies_policy_link}{The url redirecting to the cookie policy.}

\item{privacy_policy_link}{The url redirecting to the privacy policy.}

\item{legal_notice_link}{The url redirecting to the legal notice.}

\item{feedback_email}{The email address to send the feedback to.}

\item{feedback_email_subject}{The text to automatically fill the subject
field for the feedback email.}
}
\value{
Save the new footer.html in the user memory.
}
\description{
Update the footer of the dashboard, appearing in the bottom of the dashboard
}
