% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_forecasts.R
\name{update_forecasts}
\alias{update_forecasts}
\title{Update the forecasts}
\usage{
update_forecasts(file_path)
}
\arguments{
\item{file_path}{The file path of the forecasts file.}
}
\value{
Save the new forecasts.xlsx in the user memory.
}
\description{
Update the forecasts
}
