% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_logo.R
\name{update_logo}
\alias{update_logo}
\title{Update the logo}
\usage{
update_logo(file_path = NULL, position = NULL, link = NULL)
}
\arguments{
\item{file_path}{The file path of the logo.}

\item{position}{The logo position, right or left.}

\item{link}{The url to redirect the user to, when clicking on the logo.}
}
\value{
Save the new logo in the user memory.
}
\description{
Update the logo
}
