% Generated by hand; normally via roxygen2
% Please edit documentation in R/update_menu_categories.R
\name{update_menu_categories}
\alias{update_menu_categories}
\title{Update the primary menu categories}
\usage{
update_menu_categories(menu_categories)
}
\arguments{
  \item{menu_categories}{A character vector with the primary menu categories.}
}
\value{Change the primary menu categories in the configuration file.}
\description{
Update the primary menu categories used by \code{menu_categories()} for future runs.
}
