% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_meta_description.R
\name{update_meta_description}
\alias{update_meta_description}
\title{Update the meta description of the dashboard}
\usage{
update_meta_description(description)
}
\arguments{
\item{description}{The meta description.}
}
\value{
Change the meta description in the configuration file.
}
\description{
Update the meta description of the dashboard
}
