% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_overview.R
\name{update_overview}
\alias{update_overview}
\title{Update the overview section.}
\usage{
update_overview(
  overview_text = NULL,
  latest_changes_summary = NULL,
  valuebox1_value = NULL,
  valuebox1_text = NULL,
  valuebox1_icon = NULL,
  valuebox2_value = NULL,
  valuebox2_text = NULL,
  valuebox2_icon = NULL,
  valuebox3_value = NULL,
  valuebox3_text = NULL,
  valuebox3_icon = NULL
)
}
\arguments{
\item{overview_text}{The overview text.}

\item{latest_changes_summary}{The latest changes summary text.}

\item{valuebox1_value}{The value shown in the first value box.}

\item{valuebox1_text}{The text shown in the first value box.}

\item{valuebox1_icon}{The icon shown in the first value box.}

\item{valuebox2_value}{The value shown in the second value box.}

\item{valuebox2_text}{The text shown in the second value box.}

\item{valuebox2_icon}{The icon shown in the second value box.}

\item{valuebox3_value}{The value shown in the third value box.}

\item{valuebox3_text}{The text shown in the third value box.}

\item{valuebox3_icon}{The icon shown in the third value box.}
}
\value{
Change the overview information in the configuration file.
}
\description{
Update the overview section.
}
