% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_scenario.R
\name{update_scenario}
\alias{update_scenario}
\title{Update the scenario, choosing between "eurostat" and "nsi".}
\usage{
update_scenario(entity)
}
\arguments{
\item{entity}{"eurostat" or "nsi"}
}
\value{
Change the entity in the configuration file.
}
\description{
Update the scenario, choosing between "eurostat" and "nsi".
}
