% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_secondary_menu.R
\name{update_secondary_menu}
\alias{update_secondary_menu}
\title{Update the secondary menu, choosing between "storyboard" and "dropdown".}
\usage{
update_secondary_menu(type)
}
\arguments{
\item{type}{"storyboard" or "dropdown"}
}
\value{
Change the secondary menu type in the configuration file.
}
\description{
Update the secondary menu, choosing between "storyboard" and "dropdown".
}
