% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_social.R
\name{update_social}
\alias{update_social}
\title{Update the social options for the dashboard, appearing in the bottom right
corner of the navbar}
\usage{
update_social(
  article_text = NULL,
  article_link = NULL,
  facebook_link = NULL,
  twitter_link = NULL,
  linkedin_link = NULL
)
}
\arguments{
\item{article_text}{The text to show for the button redirecting to an
external article.}

\item{article_link}{The url redirecting to an external article.}

\item{facebook_link}{The url redirecting to share the page on facebook.}

\item{twitter_link}{The url redirecting to share the page on twitter.}

\item{linkedin_link}{The url redirecting to share the page on linkedin.}
}
\value{
Save the new social.html in the user memory.
}
\description{
The links for sharing the page on the social networks can be easily
created with the application https://www.sharelinkgenerator.com/.
}
