% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_title.R
\name{update_title}
\alias{update_title}
\title{Update the title of the dashboard}
\usage{
update_title(title)
}
\arguments{
\item{title}{The dashboard title.}
}
\value{
Change the title in the configuration file.
}
\description{
Update the title of the dashboard
}
