% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dashboard.R
\name{generate_dashboard}
\alias{generate_dashboard}
\title{Generate the dashboard}
\usage{
generate_dashboard(output_path, pdf_reports = FALSE)
}
\arguments{
\item{output_path}{The path where to save the html.}

\item{pdf_reports}{Enable generation of pdf reports.}
}
\value{
A html.
}
\description{
Generate the dashboard
}
